kobo/rcS: remove check for OTG kernel
With a growing number of supported Kobo models and corresponsding OTG
kernels, it becomes more difficult to properly select the correct
non-OTG kernel in rcS. With the latest change to KoboMenu, the button to
enable Nickel is disabled when running the OTG kernel. And when the user
explicitly requested to start Nickel by creating
XCSoarData/kobo/start_nickel file manually, we just respect the user's
decision and launch Nickel, even if the OTG kernel is running.