Device/Port/TTYPort: Win32 compatiblity
As TTYPort uses the boost::asio::serial_port framework, which is Win32
compatible, make the TTYPort class work on Win32.

A few methods (e. g. Flush() and Drain() cannot be implemented using
boost::asio::serial_port, so we still have platform-specific
implementations for them. For now, this includes the Write()
implementation for synchronous writing. The the Win32 code for these
platform-specific code fragments was taken from the SerialPort class.