Data/graphics/logo: workaround for cropped logo
With some tools (including recent versions of rsvg-convert, which we
use), the shadow of our logo was cropped at the edge of the rendered
image.
To solve this, these the logo was converted to Inkscape 0.92.1 format
(wich changes some meta data, chages the format of some attributes, and
introduces a viewBox), and increased the viewBox width and height by 5
px to the bottom and the right, while preserving the image width.