Thread/{Mutex,Cond}: use std::mutex and std::condition_variable if available
Now that our PosixMutex/PosixCond interfaces have been migrated to
imitate the standard library APIs, we can easily switch to the
standard library.  This avoids duplicate code.

On Windows, we keep using our own implementations, because GCC
implements std::mutex and std::condition_variable with pthread
emulation, which is not a good choice.